Refactor ResourceService and update type casts
Moved ResourceService method implementations to a new resourceService.cpp file and updated project files accordingly. Cleaned up includes in llamaModelData files and resourceService.h, and added missing includes in resourceLoader.h and audioComponent.h. Replaced C-style casts with static_cast or explicit casts in llama-adapter.h, llama-model-loader.h, and BRAIN.cpp. Removed Scene::TempLoad and related code, and improved error logging in sceneManager.cpp. Enabled _EDITOR in buildSetting.h.

+++ Source/GameApplication/Backends/llama/llama-adapter.h
@@ -78,7 +78,7 @@ struct llama_adapter_lora {
     llama_adapter_lora_weight * get_weight(ggml_tensor * w);
 
     uint32_t get_n_nodes() const {
-        return ab_map.size() * 6u; // a, b, scale, add, 2 x mul_mat
+        return (uint32_t)ab_map.size() * 6u; // a, b, scale, add, 2 x mul_mat
     }
 };
